requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
openpyxl==3.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
